Simplify MappedFileReader, close Unix fd early
Remove most of the members and make them local variables in the map()
function which is the only place that uses them.  This simplifies the
c'tor and means it cannot throw IOException anymore.

It also means we close the Unix file descriptor used to create the
mapping much earlier.  This will help reduce the load on file
descriptors.