Protection against data corruption.
If index or data files are somehow corrupted, or mismatched (meaning
that the index file from one store/version gets used in combination
with a data file for another store/version), then the server gets into
bad buffer allocation problems, causing IllegalArgumentExceptions and
potentially OOMing in the process. This commit makes it easier to
debug the issues and protect against the OOMs. Changes include:

- Disallow allocation of negative size buffers.
- Disallow allocation of excessively large value buffers (configurable,
  max 25 MB by default).
- Disallow allocation of a buffer which would go past the data file's
  length.
- Server-side logs now print out the file name and key (in hex) which
  triggered the problem.
- Server-side logs now print out the socket name and stacktrace when
  SelectorManagerWorker catches a Throwable.
- Client-side will report an "internal server error" when any of the
  bad allocation scenarios described above happen.
- Slight optimization in the allocation strategy while reading the
  key/values in a data file.